Extract PoolState and other types to a separate file
