+++ packages/hardhat-core/src/internal/hardhat-network/provider/utils/retrieveNonce.ts
@@ -1,6 +1,6 @@
 import { BN, toBuffer } from "ethereumjs-util";
 
-import { SerializedTransaction } from "../TransactionPool";
+import { SerializedTransaction } from "../PoolState";
 
 export function retrieveNonce(tx: SerializedTransaction) {
   return new BN(toBuffer(tx.get(0)));
